+++ webapp/javascript/components/TimelineChart/Tooltip.plugin.tsx
@@ -1,21 +1,24 @@
-/* eslint-disable 
-@typescript-eslint/no-unsafe-member-access, 
-@typescript-eslint/no-unsafe-call, 
-func-names, 
-@typescript-eslint/no-unsafe-return, 
-@typescript-eslint/no-unsafe-assignment,
-no-undef,
-prefer-destructuring
-*/
 import React from 'react';
 import * as ReactDOM from 'react-dom';
-import type { ExploreTooltipProps } from '@webapp/components/TimelineChart/ExploreTooltip';
 import getFormatLabel from './getFormatLabel';
 import clamp from './clamp';
 import injectTooltip from './injectTooltip';
 import { ITooltipWrapperProps } from './TooltipWrapper';
 
 const TOOLTIP_WRAPPER_ID = 'explore_tooltip_parent';
+
+// TooltipCallbackProps refers to the data available for the tooltip body construction
+export interface TooltipCallbackProps {
+  timeLabel: string;
+  values: Array<{
+    closest: number[];
+    color: number[];
+    // TODO: remove this
+    tagName: string;
+  }>;
+  coordsToCanvasPos?: jquery.flot.axis['p2c'];
+  canvasX?: number;
+}
 
 (function ($: JQueryStatic) {
   function init(plot: jquery.flot.plot & jquery.flot.plotOptions) {
@@ -63,16 +66,16 @@ const TOOLTIP_WRAPPER_ID = 'explore_tooltip_parent';
     plot.hooks!.drawOverlay!.push(() => {
       const options = plot.getOptions() as jquery.flot.plotOptions & {
         onHoverDisplayTooltip?: (
-          data: Omit<ITooltipWrapperProps & ExploreTooltipProps, 'children'>
+          data: Omit<ITooltipWrapperProps & TooltipCallbackProps, 'children'>
         ) => React.ReactElement;
       };
-      const onHoverDisplayTooltip = options.onHoverDisplayTooltip;
+      const { onHoverDisplayTooltip } = options;
       const { xaxis } = plot.getAxes() as ShamefulAny;
       const data = plot.getData();
 
       if (onHoverDisplayTooltip && exploreTooltip?.length) {
         const align = params.canvasX > plot.width() / 2 ? 'left' : 'right';
-        const timezone = options.xaxis!.timezone;
+        const { timezone } = options.xaxis!;
 
         const timeLabel = getFormatLabel({
           date: params.xToTime,
@@ -83,6 +86,7 @@ const TOOLTIP_WRAPPER_ID = 'explore_tooltip_parent';
         const values = data?.map((dataSeries, i) => {
           // Sometimes we also pass a tagName/color
           // Eg in tagExplorer page
+          // TODO: use generics
           const d = dataSeries as jquery.flot.dataSeries & {
             tagName: string;
             color: { color: number[] };