+++ webapp/javascript/pages/TagExplorerView.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useMemo } from 'react';
 import { NavLink, useLocation } from 'react-router-dom';
 import type { Maybe } from 'true-myth';
-import type { ClickEvent } from '@szhsin/react-menu';
+import type { ClickEvent } from '@webapp/ui/Menu';
 import Color from 'color';
 import TotalSamplesChart from '@webapp/pages/tagExplorer/components/TotalSamplesChart';
 import type { Profile } from '@pyroscope/models/src';
@@ -396,7 +396,9 @@ function TagExplorerView() {
         <Box>
           <div className={styles.flamegraphWrapper}>
             {type === 'loading' ? (
-              <LoadingSpinner />
+              <div className={styles.loaderContainer}>
+                <LoadingSpinner />
+              </div>
             ) : (
               <FlamegraphRenderer
                 showCredit={false}